requirements.txt
# ============================================================
# Python requirements for paiw (Pitz' AI Workshop)
# Installed inside the Docker image after PyTorch.
# ============================================================

# ── HuggingFace ecosystem ─────────────────────────────────────
transformers>=4.40.0
accelerate>=0.29.0
peft>=0.10.0
datasets>=2.19.0
huggingface-hub>=0.22.0
einops>=0.7.0
optimum>=1.19.0
# Quantization (4-bit / 8-bit) — requires CUDA on the host
bitsandbytes>=0.43.0

# ── Embeddings ───────────────────────────────────────────────
sentence-transformers>=2.7.0
FlagEmbedding>=1.2.0

# ── Inference servers / helpers ──────────────────────────────
ollama>=0.1.8

# ── RAG — LangChain stack ────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-huggingface>=0.0.3

# ── RAG — LlamaIndex stack ───────────────────────────────────
llama-index>=0.10.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-huggingface>=0.2.0

# ── Vector stores ────────────────────────────────────────────
chromadb>=0.5.0
faiss-cpu>=1.8.0
qdrant-client>=1.9.0

# ── Document / data loaders ──────────────────────────────────
pypdf>=4.2.0
python-docx>=1.1.2
beautifulsoup4>=4.12.3
html2text>=2024.2.26
unstructured>=0.13.0

# ── API clients ──────────────────────────────────────────────
openai>=1.23.0
anthropic>=0.26.0

# ── Science stack ────────────────────────────────────────────
numpy>=1.26.0
scipy>=1.13.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.2
scikit-learn>=1.4.0
pillow>=10.3.0

# ── Utilities ────────────────────────────────────────────────
tiktoken>=0.7.0
tqdm>=4.66.2
python-dotenv>=1.0.1
pyyaml>=6.0.1
requests>=2.31.0
httpx>=0.27.0
rich>=13.7.0
typer>=0.12.0

# ── Jupyter ──────────────────────────────────────────────────
jupyterlab>=4.1.0
notebook>=7.1.0
ipywidgets>=8.1.2
ipython>=8.23.0
